Add files via upload
Objective- To compare prices of stationery items on certain e-commerce websites and determine which is the most pocket friendly.

Important- The  data in this Excel spreadsheet has been created by self.

Exploring in MS-Excel- 
1. Function Shortcuts
2. Aggregate Formulas
3. Sorting 
4. Filtering
5. Customizing
6. Formatting

Analysis- The sheet compares various stationery items used in day to day life for online purchase. Depending on the quantity required, the cost consideration at Amazon, Flipkart and Snapdeal are observed. Aggregate calculations are made and the data is visually represented on a scatter chart.

Conclusion- 
1. The Title is easily visible.
2. Functions used can be observed by clicking on respective cells.
3. The chart displays which among the 3 websites should be chosen with clear representation.